+++ packages/lib/src/ref/customRef.ts
@@ -1,3 +1,4 @@
+import { action } from "mobx"
 import {
   getModelRefId,
   internalCustomRef,
@@ -45,11 +46,16 @@ export interface CustomRefOptions<T extends object> {
  * @param options Custom reference options.
  * @returns A function that allows you to construct that type of custom reference.
  */
-export function customRef<T extends object>(
-  modelTypeId: string,
-  options: CustomRefOptions<T>
-): RefConstructor<T> {
-  const getId = options.getId || getModelRefId
+export const customRef = action(
+  "customRef",
+  <T extends object>(modelTypeId: string, options: CustomRefOptions<T>): RefConstructor<T> => {
+    const getId = options.getId || getModelRefId
 
-  return internalCustomRef(modelTypeId, () => options.resolve, getId, options.onResolvedValueChange)
-}
+    return internalCustomRef(
+      modelTypeId,
+      () => options.resolve,
+      getId,
+      options.onResolvedValueChange
+    )
+  }
+)