+++ packages/lib/src/ref/index.ts
@@ -1,4 +1,10 @@
-export { getModelRefId, RefIdResolver, RefOnResolvedValueChange, RefResolver } from "./core"
+export {
+  getModelRefId,
+  getRefsResolvingTo,
+  RefIdResolver,
+  RefOnResolvedValueChange,
+  RefResolver,
+} from "./core"
 export { customRef, CustomRefOptions } from "./customRef"
-export { Ref, RefConstructor } from "./Ref"
+export { isRefOfType, Ref, RefConstructor } from "./Ref"
 export { rootRef, RootRefOptions } from "./rootRef"